add.c

+++ add.c
@@ -1,5 +1,5 @@
 #include<stdio.h>
-int rev(int n)
+long int rev(long int n)
 {
 if(n==0)
 return 0;
@@ -7,5 +7,5 @@ return 1+rev(n/10);
 }
 int main()
 {
-printf("%d\n",rev(24));
+printf("%ld\n",rev(24));
 }
